### 2023-12-04 7a0456a
DOC: Fix missing DOI:  https://github.com/openjournals/joss-reviews/issues/6117#issuecomment-1838475034
