+++ paper/paper.bib
@@ -7,14 +7,19 @@
     NUMBER = {33},
     DOI = {10.12688/f1000research.29032.1},
 }
-@Article{snaketool,
-  title   = "{Ten simple rules and a template for creating workflows-as-applications}",
-  author  = {Roach, Michael J and Pierce-Ward, N Tessa and Suchecki, Radoslaw and Mallawaarachchi, Vijini and Papudeshi, Bhavya and Handley, Scott A and Brown, C Titus and Watson-Haigh, Nathan S and Edwards, Robert A},
-  journal = {PLOS Computational Biology},
-  volume  =  {18},
-  number  =  {12},
-  pages   = {e1010705},
-  year    =  {2022}
+@article{snaketool,
+    doi = {10.1371/journal.pcbi.1010705},
+    author = {Roach, Michael J AND Pierce-Ward, N Tessa AND Suchecki, Radoslaw AND Mallawaarachchi, Vijini AND Papudeshi, Bhavya AND Handley, Scott A AND Brown, C Titus AND Watson-Haigh, Nathan S AND Edwards, Robert A},
+    journal = {PLOS Computational Biology},
+    publisher = {Public Library of Science},
+    title = {Ten simple rules and a template for creating workflows-as-applications},
+    year = {2022},
+    month = {12},
+    volume = {18},
+    url = {https://doi.org/10.1371/journal.pcbi.1010705},
+    pages = {1-9},
+    abstract = {null},
+    number = {12},
 }
 @article{minimap,
     author = {Li, Heng},
