### 2023-12-04 #6117 @beardymcjohnface
> **Author instructions**
> @beardymcjohnface if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.

Some potential reviewers matching Python and workflow/pipeline that I believe would be suitable.
- rcannood
- lparsons
- telatin
- camillescott
- ManavalanG
- jmcbroome
