### 2024-01-12 #6117 @csoneson
👋🏻 @camillescott, @lparsons, @jmcbroome - would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

> Koverage: Read-coverage analysis for massive (meta)genomics datasets
https://github.com/openjournals/joss-reviews/issues/6117

Thanks in advance!
