### 2024-01-12 #6117 @lparsons
Sure, I would be happy to help out JOSS.

On Fri, Jan 12, 2024 at 11:24 AM Charlotte Soneson ***@***.***>
wrote:

> 👋🏻 @camillescott <https://github.com/camillescott>, @lparsons
> <https://github.com/lparsons>, @jmcbroome <https://github.com/jmcbroome>
> - would you be interested in reviewing this submission for the Journal of
> Open Source Software (JOSS)?
>
> Koverage: Read-coverage analysis for massive (meta)genomics datasets
> #6117 <https://github.com/openjournals/joss-reviews/issues/6117>
>
> Thanks in advance!
>
