### 2024-01-16 #6117 @csoneson
👋🏻 @ManavalanG, @telatin - would one of you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

> Koverage: Read-coverage analysis for massive (meta)genomics datasets
https://github.com/openjournals/joss-reviews/issues/6117

Thanks in advance!
