### 2024-01-17 #6235 @editorialbot
[REVIEW]: Koverage: Read-coverage analysis for massive (meta)genomics datasets
**Submitting author:** @beardymcjohnface (<a href="http://orcid.org/0000-0003-1488-5148">Michael J. Roach</a>)
**Repository:** https://github.com/beardymcjohnface/Koverage
**Branch with paper.md** (empty if default branch): JOSS
**Version:** 0.1.11
