### 2024-01-23 102732c
BUG: New snakemake doesn't like pyhon < 3.11
