### 2024-01-23 8316de4
BUG: more dependency fixes, python 3.10 should work now
