+++ setup.py
@@ -23,6 +23,7 @@ CLASSIFIERS = [
     "Natural Language :: English",
     "Operating System :: POSIX :: Linux",
     "Operating System :: MacOS :: MacOS X",
+    "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
     "Topic :: Scientific/Engineering :: Bio-Informatics",
@@ -42,7 +43,7 @@ setup(
     name="koverage",
     packages=packages,
     url="https://github.com/beardymcjohnface/Koverage",
-    python_requires=">=3.11",
+    python_requires=">=3.10",
     description="Quickly get coverage statistics given reads and an assembly",
     long_description=get_description(),
     long_description_content_type="text/markdown",
@@ -55,6 +56,7 @@ setup(
         "snaketool-utils>=0.0.4",
         "metasnek>=0.0.8",
         "snakemake>=7.14.0",
+        "pulp<2.8",
         "pyyaml>=6.0",
         "Click>=8.1.3",
         "zstandard>=0.21.0",
