+++ .github/workflows/build-koverage-envs.yaml
@@ -65,4 +65,4 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           python -m pip install .
-          koverage test map kmer coverm --conda-create-envs-only
+          koverage test build_envs
