### 2024-01-27 beardymcjohnface/Koverage#36 @telatin
JOSS Review
Just a tracker for the JOSS review

1) Installation help #35
2) Typo #39

See https://github.com/openjournals/joss-reviews/issues/6235#issuecomment-1913178466
