+++ koverage/scripts/koverageReport.py
@@ -1,5 +1,3 @@
-import os
-import yaml
 import pandas as pd
 import datapane as dp
 from plotly.subplots import make_subplots
@@ -130,18 +128,18 @@ def generate_figure(all_df, ref_fa, buttons):
 
 
 def main(**kwargs):
-    if kwargs["pyspy"]:
-        subprocess.Popen(
-            [
-                "py-spy",
-                "record",
-                "-s",
-                "-o",
-                kwargs["pyspy_svg"],
-                "--pid",
-                str(os.getpid()),
-            ]
-        )
+    # if kwargs["pyspy"]:
+    #     subprocess.Popen(
+    #         [
+    #             "py-spy",
+    #             "record",
+    #             "-s",
+    #             "-o",
+    #             kwargs["pyspy_svg"],
+    #             "--pid",
+    #             str(os.getpid()),
+    #         ]
+    #     )
     # Read in data frames
     sample_df = pd.read_csv(kwargs["sample_cov"], sep="\t")
     all_df = pd.read_csv(kwargs["all_cov"], sep="\t")
@@ -193,6 +191,6 @@ if __name__ == "__main__":
         sample_names=snakemake.params.sample_names,
         ref_fasta=snakemake.params.ref_fasta,
         max_ctg=snakemake.params.max_ctg,
-        pyspy=snakemake.params.pyspy,
-        pyspy_svg=snakemake.log.pyspy,
+        # pyspy=snakemake.params.pyspy,
+        # pyspy_svg=snakemake.log.pyspy,
     )
