### 2024-01-29 563dbf8
BUG: jellyfish with strict channel_priority requires bioconda first
