+++ koverage/workflow/envs/jellyfish.yaml
@@ -1,7 +1,7 @@
 name: jellyfish
 channels:
-    - conda-forge
     - bioconda
+    - conda-forge
     - defaults
 dependencies:
     - jellyfish=2.2.10
