### 2024-01-29 efb2286
MAINT: test with strict channel_priority
