+++ .github/workflows/codecov.yaml
@@ -5,23 +5,25 @@ on:
   push:
     branches: [ "main" ]
     paths:
-      - '.github/**'
+      - '.github/workflows/codecov.yaml'
+      - '.github/.codecov.yml'
       - 'build/**'
       - 'config/**'
       - 'tests/**'
       - 'test_data/**'
-      - 'workflow/codecov.yaml'
+      - 'workflow/**'
       - 'koverage/**'
       - 'setup.py'
   pull_request:
     branches: [ "main" ]
     paths:
-      - '.github/**'
+      - '.github/workflows/codecov.yaml'
+      - '.github/.codecov.yml'
       - 'build/**'
       - 'config/**'
       - 'tests/**'
       - 'test_data/**'
-      - 'workflow/codecov.yaml'
+      - 'workflow/**'
       - 'koverage/**'
       - 'setup.py'
 
