+++ .github/workflows/py-app.yaml
@@ -5,23 +5,23 @@ on:
   push:
     branches: [ "main" ]
     paths:
-      - '.github/**'
+      - '.github/workflows/py-app.yaml'
       - 'build/**'
       - 'config/**'
       - 'tests/**'
       - 'test_data/**'
-      - 'workflow/py-app.yaml'
+      - 'workflow/**'
       - 'koverage/**'
       - 'setup.py'
   pull_request:
     branches: [ "main" ]
     paths:
-      - '.github/**'
+      - '.github/workflows/py-app.yaml'
       - 'build/**'
       - 'config/**'
       - 'tests/**'
       - 'test_data/**'
-      - 'workflow/py-app.yaml'
+      - 'workflow/**'
       - 'koverage/**'
       - 'setup.py'
 
