+++ .github/workflows/build-koverage-envs.yaml
@@ -5,23 +5,23 @@ on:
   push:
     branches: [ "main" ]
     paths:
-      - '.github/**'
+      - '.github/workflows/build-koverage-envs.yaml'
       - 'build/**'
       - 'config/**'
       - 'tests/**'
       - 'test_data/**'
-      - 'workflow/build-koverage-envs.yaml'
+      - 'workflow/**'
       - 'koverage/**'
       - 'setup.py'
   pull_request:
     branches: [ "main" ]
     paths:
-      - '.github/**'
+      - '.github/workflows/build-koverage-envs.yaml'
       - 'build/**'
       - 'config/**'
       - 'tests/**'
       - 'test_data/**'
-      - 'workflow/build-koverage-envs.yaml'
+      - 'workflow/**'
       - 'koverage/**'
       - 'setup.py'
 
