+++ .github/workflows/build-koverage-envs.yaml
@@ -6,7 +6,6 @@ on:
     branches: [ "main" ]
     paths:
       - '.github/workflows/build-koverage-envs.yaml'
-      - 'build/**'
       - 'config/**'
       - 'tests/**'
       - 'test_data/**'
@@ -17,7 +16,6 @@ on:
     branches: [ "main" ]
     paths:
       - '.github/workflows/build-koverage-envs.yaml'
-      - 'build/**'
       - 'config/**'
       - 'tests/**'
       - 'test_data/**'
@@ -57,7 +55,6 @@ jobs:
           mamba-version: "*"
           channels: conda-forge,bioconda,defaults
           channel-priority: strict
-          activate-environment: koverage
           python-version: ${{ matrix.python-version }}
           auto-activate-base: false
 
