### 2024-02-03 4d1e406
BUG: fix use snakemake <8
