### 2024-02-03 33b39a7
MAINT: remove f-strings for python 3.12
