### 2024-02-08 6eef5fd
MAINT: copy-paste err and test fix
