### 2024-02-16 a1ad9cd
MAINT: update install instructions to address #42
