### 2024-02-25 67ce277
Minor modifications to paper and bibliography
