+++ paper/paper.bib
@@ -9,7 +9,7 @@
 }
 @article{snaketool,
     doi = {10.1371/journal.pcbi.1010705},
-    author = {Roach, Michael J AND Pierce-Ward, N Tessa AND Suchecki, Radoslaw AND Mallawaarachchi, Vijini AND Papudeshi, Bhavya AND Handley, Scott A AND Brown, C Titus AND Watson-Haigh, Nathan S AND Edwards, Robert A},
+    author = {Roach, Michael J and Pierce-Ward, N Tessa and Suchecki, Radoslaw and Mallawaarachchi, Vijini and Papudeshi, Bhavya and Handley, Scott A and Brown, C Titus and Watson-Haigh, Nathan S and Edwards, Robert A},
     journal = {PLOS Computational Biology},
     publisher = {Public Library of Science},
     title = {Ten simple rules and a template for creating workflows-as-applications},
@@ -147,7 +147,7 @@
   howpublished = {\url{https://github.com/beardymcjohnface/metasnek}},
 }
 @article{coral,
-	author = {Lima, Laís FO and Alker, Amanda T and Papudeshi, Bhavya and Morris, Megan M and Edwards, Robert A and de Putron, Samantha J and Dinsdale, Elizabeth A},
+	author = {Lima, Laís FO and Alker, Amanda T and Papudeshi, Bhavya and Morris, Megan M and Edwards, Robert A and {de Putron}, Samantha J and Dinsdale, Elizabeth A},
 	title = "{Coral and Seawater Metagenomes Reveal Key Microbial Functions to Coral Health and Ecosystem Functioning Shaped at Reef Scale}",
 	year = {2023},
 	doi = {10.1007/s00248-022-02094-6},
