### 2024-02-25 #6235 @csoneson
👋🏻 @beardymcjohnface - I have gone through your submission and overall it looks good. I made a PR [here](https://github.com/beardymcjohnface/Koverage/pull/44) with some small suggestions for the paper.

In addition to that, note that your paper is currently significantly longer than the recommendations in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) (250-1,000 words, yours is over 2,000). If you could look into trimming the manuscript a bit to get closer to the recommended length, that would be great. Once that is done, you can generate a new proof here with `@editorialbot generate pdf`.

I will make another post below with the other remaining steps (some for you, some for me).
