+++ paper/paper.md
@@ -10,19 +10,19 @@ authors:
     orcid: 0000-0003-1488-5148
     affiliation: "1,2"
     corresponding: true
-  - name: Brad Hart
+  - name: Bradley J. Hart
     orcid: 0000-0001-8110-2460
     affiliation: 3
-  - name: Sarah Beecroft
+  - name: Sarah J. Beecroft
     orcid: 0000-0002-3935-2279
     affiliation: 4
   - name: Bhavya Papudeshi
     orcid: 0000-0001-5359-3100
     affiliation: 1
-  - name: Laura Inglis
+  - name: Laura K. Inglis
     orcid: 0000-0001-7919-8563
     affiliation: 1
-  - name: Susanna Grigson
+  - name: Susanna R. Grigson
     orcid: 0000-0003-4738-3451
     affiliation: 1
   - name: Vijini Mallawaarachchi
@@ -53,27 +53,27 @@ bibliography: paper.bib
 
 # Summary
 
-Genomes of organisms are constructed by assembling short fragments (called sequencing reads) that are the resulting data outputs of whole genome sequencing (WGS). It is useful to determine the read-coverage of sequencing reads in the resulting genome assembly for many reasons, such as identifying duplication or deletion events, identifying related contigs for binning in metagenome assemblies [@metacoag;@graphbin2], or analysing taxonomic compositions of metagenomic samples [@condiga]. Although calculating the read-coverage of sequencing reads to a reference genome is a routine task, it typically involves several read and write operations (I/O operations) of the sequencing data. Although this is not a problem for small datasets, it can be a significant bottleneck when analysing a large number of samples, or when screening very large reference sequence files. Koverage is designed to reduce the I/O burden as much as possible to enable maximum scalability for large sample sizes. Koverage also includes a kmer-based coverage method that significantly reduces the computational complexity of screening large reference genomes such as the human genome. Koverage is a Snakemake [@snakemake] based pipeline, providing out-of-the-box support for HPC and cloud environments. It utilises the Snaketool [@snaketool] command line interface and is available to install via PIP or Conda for maximum ease of use. The source code and documentation is available at [https://github.com/beardymcjohnface/Koverage](https://github.com/beardymcjohnface/Koverage).
+Genomes of organisms are constructed by assembling sequencing from whole genome sequencing (WGS). It is useful to determine sequencing read-coverage of the genome assembly, for instance identifying duplication or deletion events, related contigs for binning metagenomes [@metacoag;@graphbin2], or analysing taxonomic compositions of metagenomes [@condiga]. Although calculating read-coverage is a routine task, it typically involves several complete read and write operations (I/O operations). This is not a problem for small datasets, but can be a significant bottleneck for very large datasets. Koverage reduces I/O burden as much as possible to enable maximum scalability. Koverage includes a kmer-based method that significantly reduces the computational complexity for very large reference genomes. Koverage uses Snakemake [@snakemake], providing out-of-the-box support for HPC and cloud environments. It utilises the Snaketool [@snaketool] command line interface, and is installable with PIP or Conda for maximum ease of use. Source code and documentation are available at [https://github.com/beardymcjohnface/Koverage](https://github.com/beardymcjohnface/Koverage).
 
 
 # Statement of need
 
[…]
