### 2024-02-27 #6235 @Kevin-Mattheus-Moerman
@beardymcjohnface as AEiC will now process final steps for acceptance in JOSS. I have just checked the paper, the archive link, your repository, and this review. All seems in order so I will now proceed to accept this paper in JOSS.
